requirements.txt
flask
flask-cors
numpy
tensorflow
ultralytics
Pillow
joblib
opencv-python
scikit-learn
pymongo
python-dotenv
inference-sdk